Moved GenericMetricController.makeBarChartConfig() to HgraphCartController.
