+++ src/charts/hgraph/GenericMetricController.ts
@@ -69,76 +69,4 @@ export abstract class GenericMetricController extends HgraphChartController {
     }
 
 
-    //
-    // Protected (tools for subclasses)
-    //
-
-    protected makeBarChartConfig(metrics: EcosystemMetric[], range: ChartRange,
-                                 logarithmic: boolean, yLabel: string|null): ChartConfiguration {
-        const granularity = computeGranularityForRange(range)
-        const graphLabels = makeGraphLabels(metrics, granularity)
-        const graphDataSet = this.makeGraphDataSet(metrics) as any
-        const textPrimaryColor = this.themeController.getTextPrimaryColor()
-        const textSecondaryColor = this.themeController.getTextSecondaryColor()
-        const yScaleType = logarithmic ? "logarithmic" : "linear"
-
-        return {
-            type: 'bar',
-            data: {
-                labels: graphLabels,
-                datasets: [graphDataSet],
-            },
-            options: {
-                plugins: {
-                    legend: {
-                        display: false
-                    }
-                },
-                scales: {
-                    x: {
-                        ticks: {
-                            color: textPrimaryColor
-                        },
-                        grid: {
-                            display: false
-                        }
-                    },
-                    y: {
-                        type: yScaleType,
-                        ticks: {
-                            color: textPrimaryColor
-                        },
-                        grid: {
-                            display: false
-                        },
-                        beginAtZero: true,
-                        title: {
-                            display: yLabel !== null,
-                            text: yLabel ?? "",
-                            color: textSecondaryColor
-                        },
-                    }
-                },
-                maintainAspectRatio: false
-            }
-        }
-    }
-
-
-    //
-    // Private
-    //
-
-    protected makeGraphDataSet(metrics: EcosystemMetric[]): object {
-        const totals = metrics.map((m: EcosystemMetric) => m.total)
-        const graphBarColor = this.themeController.getGraphBarColor()
-        return {
-            label: this.chartTitle,
-            data: totals,
-            borderWidth: 1,
-            backgroundColor: graphBarColor
-        }
-    }
-
-
 }
